Cleanup state folder, broke models into their own files; Updated imports
