+++ pages/eligibility/index.tsx
@@ -10,22 +10,16 @@ import { ConditionalLinks } from '../../components/ConditionalLinks'
 import { ContactCTA } from '../../components/ContactCTA'
 import { FAQ } from '../../components/FAQ'
 import { ComponentFactory } from '../../components/Forms/ComponentFactory'
-import { Input } from '../../components/Forms/Input'
 import { useStore } from '../../components/Hooks'
 import { Layout } from '../../components/Layout'
 import { NeedHelpList } from '../../components/Layout/NeedHelpList'
 import ProgressBar from '../../components/ProgressBar'
 import { Tooltip } from '../../components/Tooltip/tooltip'
-import {
-  EstimationSummaryState,
-  ResultKey,
-} from '../../utils/api/definitions/enums'
-import { BenefitResult } from '../../utils/api/definitions/types'
+import { EstimationSummaryState } from '../../utils/api/definitions/enums'
 import { validateIncome } from '../../utils/api/helpers/validator'
 import { dataFetcher } from '../../utils/web/helpers/utils'
 import { Instance } from 'mobx-state-tree'
-import { RootStore, Form } from '../../client-state/store'
-import { linearBuckets } from 'prom-client'
+import { RootStore } from '../../client-state/store'
 import { observer } from 'mobx-react'
 import { ResultsTable } from '../../components/ResultsTable'
